pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyfil"
version = "0.0.1"
authors = [
  { name="Daniel Henrique da Silva",   email="danielhenrique@edu.univali.br" },
  { name="Guilherme Melo",             email="@edu.univali.br" },
  { name="Leonardo Pinheiro de Souza", email="@edu.univali.br" },
]
description = "Biblioteca Python para processamento de imagens e aplicação de filtros. Desenvolvida durante disciplina Introdução à Python."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "Pillow",
    "numpy",
]

[project.urls]
"Homepage" = "https://github.com/danielhenriquee/pyfil-library"